% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTaxonKeyCounts.R
\name{getTaxonKeyCounts}
\alias{getTaxonKeyCounts}
\title{Get occ counts by taxonKey and datasetkey.}
\usage{
getTaxonKeyCounts(datasetkey, Step = 1000, maxPages = 100,
  verbose = TRUE)
}
\arguments{
\item{datasetkey}{gbif datasetkey uuid string.}

\item{Step}{The step size. Something small like 100, 1000.}

\item{maxPages}{The max number of pages to go through.}

\item{verbose}{logical indicating printing or not.}
}
\value{
Returns data.frame of results.
}
\description{
Get occ counts by taxonKey and datasetkey.
}
\examples{

\dontrun{

getTaxonKeyCounts(datasetkey = "181eab51-9399-4baa-a0df-8de01a3acf19",Step=1000,maxPages=100,verbose=TRUE)

}

}
