% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pager.R
\name{pager}
\alias{pager}
\title{Page through api calls.}
\usage{
pager(Step, Page)
}
\arguments{
\item{Step}{The step size. Something like 100, 1000, ect. Not 10,000.}

\item{Page}{The page number.}
}
\value{
A string of "&facetLimit=10&facetOffset=1000" to stick on the end of an api call.
}
\description{
Page through api calls.
}
\examples{

\dontrun{
gbifapi::gbifapi(prefix \%+\% "&" \%+\% "datasetKey=" \%+\% datasetKey \%+\% "&" \%+\% pager(Step,Page))
}

}
