% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageWhile.R
\name{pageWhile}
\alias{pageWhile}
\title{Utility function to page through results until done.}
\usage{
pageWhile(FUN = f, Step = 100, maxPages = 10, verbose = TRUE)
}
\arguments{
\item{FUN}{anonymous function with api call to pass.}

\item{Step}{numeric step size.}

\item{maxPages}{numeric maximum number of pages to go through.}

\item{verbose}{logical indicating whether to print page output.}
}
\value{
a list all the pages called.
}
\description{
Utility function to page through results until done.
}
\examples{

\dontrun{


key = datasetkey #

f = function(datasetkey=key, ...) { # anonymous function to be run with page while
 x = list(...) # get Step and Page variables passed in by pageWhile
 prefix = "http://api.gbif.org/v1/occurrence/search?limit=0&facet=taxonKey"
 gbifapi::gbifapi(prefix \%+\% "&" \%+\% "datasetKey=" \%+\% datasetkey \%+\% "&" \%+\% gbifapi::pager(x$Step,x$Page))$facets[[1]]$counts
}

CL = pageWhile(FUN=f,Step=100,maxPages=1) # get pages while is not done.





}

}
